Refactoring, restricting access in IndyWallet
